TST: Test invalid directive not at top
