+++ openff/interchange/_tests/data/invalid_partway_through.top
@@ -1,0 +1,35 @@
+[ defaults ]
+; nbfunc        comb-rule       gen-pairs       fudgeLJ fudgeQQ
+     1 2      yes    0.50000000 0.83333333
+
+[ atomtypes ]
+;type, bondingtype, atomic_number, mass, charge, ptype, sigma, epsilon
+Cl-         Cl-       17        35.45000000         0.00000000 A        4.47765700e-01     1.48912740e-01
+HW          HW         1         1.00800000         0.00000000 A        0.00000000e+00     0.00000000e+00
+Na+         Na+       11        22.99000000         0.00000000 A        2.43928070e-01     3.65846030e-01
+OW          OW         8        16.00000000         0.00000000 A        3.15075240e-01     6.35968000e-01
+c           c          6        12.01000000         0.00000000 A        3.39966950e-01     3.59824000e-01
+c2          c2         6        12.01000000         0.00000000 A        3.39966950e-01     3.59824000e-01
+c3          c3         6        12.01000000         0.00000000 A        3.39966950e-01     4.57729600e-01
+ca          ca         6        12.01000000         0.00000000 A        3.39966950e-01     3.59824000e-01
+h1          h1         1         1.00800000         0.00000000 A        2.47135300e-01     6.56888000e-02
+h2          h2         1         1.00800000         0.00000000 A        2.29317330e-01     6.56888000e-02
+h4          h4         1         1.00800000         0.00000000 A        2.51055260e-01     6.27600000e-02
+ha          ha         1         1.00800000         0.00000000 A        2.59964250e-01     6.27600000e-02
+hc          hc         1         1.00800000         0.00000000 A        2.64953280e-01     6.56888000e-02
+hn          hn         1         1.00800000         0.00000000 A        1.06907850e-01     6.56888000e-02
+ho          ho         1         1.00800000         0.00000000 A        0.00000000e+00     0.00000000e+00
+hx          hx         1         1.00800000         0.00000000 A        1.95997720e-01     6.56888000e-02
+n           n          7        14.01000000         0.00000000 A        3.24999850e-01     7.11280000e-01
+n4          n4         7        14.01000000         0.00000000 A        3.24999850e-01     7.11280000e-01
+nb          nb         7        14.01000000         0.00000000 A        3.24999850e-01     7.11280000e-01
+o           o          8        16.00000000         0.00000000 A        2.95992190e-01     8.78640000e-01
+oh          oh         8        16.00000000         0.00000000 A        3.06647340e-01     8.80313600e-01
+os          os         8        16.00000000         0.00000000 A        3.00001230e-01     7.11280000e-01
+
+[ moleculetype ]
+HST          3
+
+[ flag ]
+a 1
+b 2
